invoice generator done

+++ backEnd/src/main/java/lk/ijse/backend/service/PlacePaymentService.java
@@ -3,5 +3,5 @@ package lk.ijse.backend.service;
 import lk.ijse.backend.DTO.PlacePaymentDTO;
 
 public interface PlacePaymentService {
-    int placePayment(PlacePaymentDTO placePaymentDTO);
+    long placePayment(PlacePaymentDTO placePaymentDTO);
 }
